Move RTL detectiong logic into react useEffect
